requirements.txt
torch
torchvision
dlib
scikit-image
easydict
PyYAML
dominate>=2.3.1
dill
tensorboardX
scipy
opencv-python
einops
PySimpleGUI
matplotlib
flask
flask-restful
python-dotenv